pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "salmopy"
version = "0.42.0"
description = "Python conversion of inSTREAM/inSALMO 7.4 individual-based salmonid model"
authors = [{name = "Salmopy Team"}]
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "geopandas>=0.14",
    "mesa>=3.1",
    "meshio>=5.3",
    "numpy>=1.24",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "scipy>=1.11",
    "shapely>=2.0",
]

[project.urls]
Homepage = "https://github.com/razinkele/instream-py"
Repository = "https://github.com/razinkele/instream-py"

[project.optional-dependencies]
numba = [
    "numba>=0.59",
]
jax = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "interpax>=0.3",
]
calibration = [
    "SALib>=1.4",
    "scikit-learn>=1.3",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "salmopy[numba]",
    "salmopy[calibration]",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
]
frontend = [
    "shiny>=1.0",
    "shiny-deckgl>=1.9",
    "plotly>=6.0",
    "shinyswatch>=0.9",
]

[project.scripts]
salmopy = "salmopy.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/salmopy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: e2e tests that drive the live Shiny app (opt-in via E2E_INTEGRATION=1)",
]

[tool.coverage.run]
source = ["salmopy"]
omit = ["*/backends/jax_backend/*", "*/backends/numba_backend/*"]

[tool.coverage.report]
fail_under = 80
